Add option to remove expired locks instead of modify them
Currenly when a lock is expired its record in MongoDB is modified
to include information about its expiry, along with changing its
name.

There may be cases where the end user will not want this to be the
case. For example, when trying to reduce the amount of data stored
in the collection itself when performing many `acquire` and `release`
actions.

This adds an option to the lock options Object called `removeExpired`.
When this is set to true, instead of performing a `findAndModify` on
the collection a `findAndRemove` is invoked instead.

This is compatible with both 1.X.X and 2.X.X streams of the MongoDB
client.

Tests have been added to test the current behaviour remains the default
and that the `removeExpired` option actually results in these records
being removed.